Yuxuan: update on PRT renderer dependency (#237)

+++ lib/renderer/prt_util.py
@@ -158,7 +158,7 @@ def computePRT(mesh_path, scale, n, order):
             hits = mesh.ray.intersects_any(origins + delta * normals, vectors)
             nohits = np.logical_and(front, np.logical_not(hits))
 
-            PRT = (nohits.astype(np.float) * dots)[:, None] * SH
+            PRT = (nohits.astype(np.float32) * dots)[:, None] * SH
 
             if PRT_all is not None:
                 PRT_all += (PRT.reshape(-1, n, SH.shape[1]).sum(1))
